+++ index.js
@@ -1,2 +1,0 @@
-export { TestUtils } from './src/test-utils';
-export { Simulate } from './src/simulate';