+++ lib/index.d.ts
@@ -1,0 +1,3 @@
+export { TestUtils } from './test-utils';
+export { Simulate } from './simulate';
+//# sourceMappingURL=index.d.ts.map