+++ src/index.ts
@@ -1,0 +1,2 @@
+export { TestUtils } from './test-utils';
+export { Simulate } from './simulate';